backend/requirements.txt
fastapi
uvicorn
apscheduler
supabase
python-dotenv
requests
pydantic
pandas
python-multipart
openpyxl
groq

# Auth dependencies
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
